Add actionTypeCreator
